Create image_processor.py
This module uses Google Vision API to detect the text in a raster map, and send the textual results to Google NLP API to identify addresses and locations. The resulting entities are then consumed by the Google Geocoding API to determine where the map belongs to.

The Vision API has a limit of 20 MB file size, and some of the maps reach up to 300 MB in size,  causing requests timeouts. This module is a starting point to optimize for this problem using a recursion. The main function  takes in the image and divide it into two parts each time, once horizontally and once vertically until the 20 MB size limit is reached.